requirements.txt
certifi==2020.12.5
chardet==3.0.4
colorzero==1.1
gpiozero==1.5.1
idna==2.10
requests==2.25.0
RPi.bme280==0.2.3
RPi.GPIO==0.7.0
smbus2==0.4.0
spidev==3.5
urllib3==1.26.2
